spawnerStatus.tsを作成する
